Adding return values

+++ include/locate.h
@@ -2,5 +2,4 @@
 #include "global.h"
 
 
-
-void locate (float x, float *xx, int *j, int n);
+void locate(double x, double *xx, int *j, int n);
