+++ include/tabfind.h
@@ -3,5 +3,4 @@
 #include "locate.h"
 
 
-
-void tabfind (double temperature, double *rate, float *atomic_temp_tab);
+void tabfind(double temperature, double *rate, double *atomic_temp_tab);
