+++ src/tabfind.cpp
@@ -2,10 +2,10 @@
 
 
 void
-tabfind(double temperature, double *rate, float *atomic_temp_tab) {
+tabfind(double temperature, double *rate, double *atomic_temp_tab) {
 
 
-    float cooling_tab[50] = {
+    double cooling_tab[50] = {
             -22.97, -22.87, -22.77, -22.64, -22.52, -22.42, -22.31, -22.22, -22.14,
             -22.07, -22.02, -21.97, -21.91, -21.86, -21.81, -21.75, -21.69, -21.63,
             -21.58, -21.53, -21.49, -21.45, -21.41, -21.38, -21.37, -21.36, -21.37,
